Adds a test_page
The test_page is gonna be filler till we get our actual main page up. This way i can get the code working without waiting for the page to be up. This pages goal is for me to be able to loop through the stimuli like it would for a patient.